Phase 2: proposals data model, approval CLI, and enforced guardrails
- Switch all monetary fields (Proposal.estimated_profit, profit-guard
  settings) from float to Decimal; DB stores them as Numeric(12,2).
- Implement guardrails for real (retail-arbitrage wording, profit
  guard, rate budget, supplier stock) with a deny-by-default policy:
  missing info or unconfirmed wholesale sourcing denies rather than
  passing. Unskips all Phase 0 guardrail test placeholders.
- Add guardrails/gateway.execute_side_effect as the single chokepoint
  for any write side effect (publish/price_change/withdraw/purchase):
  requires APPROVED status and every applicable guardrail to pass
  before invoking the executor callback. A static test asserts no
  other source file calls the EbayClient write methods outside the
  gateway and its own definition, so a future bypass gets caught.
- Add SQLAlchemy models/repository for `proposals` (DB-agnostic types
  only) plus an Alembic migration, verified by actually running
  `alembic upgrade head`/`downgrade` against SQLite in tests.
  Repository enforces pending->approved/rejected->executed/failed
  transitions and rejects anything else.
- Wire an approval CLI (`ebay-dropship proposals list/approve/reject`)
  as a console script on top of the repository.

+++ ebay-dropship-agent/src/ebay_dropship/approval/__init__.py
@@ -2,6 +2,7 @@
 
 from abc import ABC, abstractmethod
 from datetime import datetime
+from decimal import Decimal
 from enum import StrEnum
 from typing import Any
 
@@ -52,7 +53,7 @@ class Proposal(BaseModel):
     summary: str
     rationale: str
     risk_level: RiskLevel
-    estimated_profit: float | None = None
+    estimated_profit: Decimal | None = None  # 金額は Decimal 固定(float禁止)
     requires_human_approval: bool
     payload: dict[str, Any] = {}
     status: ProposalStatus = ProposalStatus.PENDING
